Staff code cleanup
