+++ registration/templates/registration/emails/new-staff.txt
