+++ registration/templates/registration/staff/staff-new-payment.html
@@ -143,7 +143,7 @@
           $.ajax({
               "type": "POST",
               "dataType": "json",
-              "url": "{% url 'registration:addNewStaff' %}",
+              "url": "{% url 'registration:add_new_staff' %}",
               "data": JSON.stringify(data),
               "beforeSend": function (xhr, settings) {
                   console.log("Before Send");
