+++ registration/templates/registration/staff/staff-new.html
@@ -72,7 +72,7 @@
           $.ajax({
               "type": "POST",
               "dataType": "json",
-              "url": "{% url 'registration:findNewStaff' %}",
+              "url": "{% url 'registration:find_new_staff' %}",
               "data": JSON.stringify(data),
               "beforeSend": function (xhr, settings) {
                   console.log("Before Send");
@@ -82,7 +82,7 @@
                   $("#errorModal").modal('show');
               },
               "success": function () {
-                  window.location = "{% url 'registration:infoNewStaff' %}";
+                  window.location = "{% url 'registration:info_new_staff' %}";
               }
           });
       }
